Rename footer folder with the name of the component
