+++ components/ui/PageFooter/FooterSection.vue
@@ -12,7 +12,7 @@
 <script lang="ts">
 import Vue from 'vue'
 import { Component, Prop } from 'vue-property-decorator'
-import FooterSubsection from '~/components/ui/footer/FooterSubsection.vue'
+import FooterSubsection from '~/components/ui/PageFooter/FooterSubsection.vue'
 
 @Component({
   components: {
